put PlannerContext into ExecutionProps
