+++ benchmarks/src/cancellation.rs
@@ -30,10 +30,10 @@ use datafusion::datasource::file_format::FileFormat;
 use datafusion::datasource::listing::{ListingOptions, ListingTableUrl};
 use datafusion::execution::object_store::ObjectStoreUrl;
 use datafusion::execution::TaskContext;
+use datafusion::logical_expr::planner_context::PlannerContext;
 use datafusion::physical_plan::coalesce_partitions::CoalescePartitionsExec;
 use datafusion::physical_plan::ExecutionPlan;
 use datafusion::prelude::*;
-use datafusion::sql::planner::PlannerContext;
 use datafusion_common::instant::Instant;
 use futures::TryStreamExt;
 use object_store::ObjectStore;
