+++ datafusion/sql/src/expr/mod.rs
@@ -19,6 +19,7 @@ use arrow::datatypes::{DataType, TimeUnit};
 use datafusion_expr::planner::{
     PlannerResult, RawBinaryExpr, RawDictionaryExpr, RawFieldAccessExpr,
 };
+use datafusion_expr::planner_context::PlannerContext;
 use sqlparser::ast::{
     AccessExpr, BinaryOperator, CastFormat, CastKind, DataType as SQLDataType,
     DictionaryField, Expr as SQLExpr, ExprWithAlias as SQLExprWithAlias, MapEntry,
@@ -37,7 +38,7 @@ use datafusion_expr::{
     Operator, TryCast,
 };
 
-use crate::planner::{ContextProvider, PlannerContext, SqlToRel};
+use crate::planner::{ContextProvider, SqlToRel};
 
 mod binary_op;
 mod function;
