+++ datafusion/sql/src/expr/substring.rs
@@ -15,10 +15,11 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::planner::{ContextProvider, PlannerContext, SqlToRel};
+use crate::planner::{ContextProvider, SqlToRel};
 use datafusion_common::{not_impl_err, plan_err};
 use datafusion_common::{DFSchema, Result, ScalarValue};
 use datafusion_expr::planner::PlannerResult;
+use datafusion_expr::planner_context::PlannerContext;
 use datafusion_expr::Expr;
 use sqlparser::ast::Expr as SQLExpr;
